pyproject.toml
[project]
name = "chatbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.29.0",
    "streamlit>=1.48.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]
